requirements.txt
# Core dependencies
discord.py-self>=2.0
aiohttp>=3.9.0

# Captcha recognition
onnxruntime>=1.17.0
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.0